requirements.txt
langchain==0.3.26
flask==3.1.0
sentence-transformers==3.4.1.0
pypdf==5.6.1
python-dotenv==1.0.1
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
-e .
